small fixes + tipo krasivo  tipo mvp
