+++ app/src/main/java/com/example/numfac/model/ApiFactory.kt
@@ -1,0 +1,19 @@
+package com.example.numfac.model
+
+import retrofit2.Retrofit
+import retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory
+import retrofit2.converter.gson.GsonConverterFactory
+
+class ApiFactory {
+    companion object {
+        fun createApi(): NumFacApiService {
+            val retrofit = Retrofit.Builder()
+                .addCallAdapterFactory(RxJava2CallAdapterFactory.create())
+                .addConverterFactory(GsonConverterFactory.create())
+                .baseUrl(URL)
+                .build()
+            return retrofit.create(NumFacApiService::class.java);
+        }
+        private const val URL = "http://numbersapi.com/"
+    }
+}
