+++ app/src/main/java/com/example/numfac/presenter/DateListPresenter.kt
@@ -3,9 +3,9 @@ package com.example.numfac.presenter
 import com.example.numfac.model.NumFacModel
 import com.example.numfac.view.fragments.DateListView
 
-open class DateListPresenter(val model: NumFacModel, val view: DateListView) {
+class DateListPresenter(private val model: NumFacModel,private val view: DateListView) {
 
-    fun setDateList() {
+    fun setDateList() =
         view.showDateList(model.getDateList())
-    }
+
 }
